VMS-11784: Fix search progress calculation
<==== Mercurial data ====>
Branch: x_VMS-11784_manual_searcher_crashes

+++ vms/server/nx_vms_server/src/core/resource_management/manual_camera_search_task_manager.cpp
@@ -102,7 +102,7 @@ double QnManualSearchTaskManager::doneToTotalTasksRatio() const
     if (totalTasks == 0)
         return 0;
 
-    return (totalTasks - m_remainingTaskCount) / totalTasks;
+    return (totalTasks - m_remainingTaskCount) / double(totalTasks);
 }
 
 QnManualResourceSearchList QnManualSearchTaskManager::foundResources() const
